+++ examples/kaggle-plasticc/kaggle_plasticc.py
@@ -1,7 +1,7 @@
 import pandas as pd
 
 from sklearn.model_selection import StratifiedKFold
-from nyaggle.experiment import experiment
+from nyaggle.experiment import run_experiment
 
 
 meta = pd.read_csv('training_set_metadata.csv')
@@ -18,9 +18,9 @@ lgb_param_extra = {
 
 skf = StratifiedKFold(n_splits=5, shuffle=True, random_state=0)
 
-result_extra = experiment(lgb_param_extra,
-                          meta_extra.drop('target', axis=1),
-                          meta_extra['target'],
-                          logging_directory='plasticc-{time}',
-                          cv=skf,
-                          type_of_target='multiclass')
+result_extra = run_experiment(lgb_param_extra,
+                              meta_extra.drop('target', axis=1),
+                              meta_extra['target'],
+                              logging_directory='plasticc-{time}',
+                              cv=skf,
+                              type_of_target='multiclass')
