+++ nyaggle/experiment/__init__.py
@@ -1,3 +1,3 @@
 from nyaggle.experiment.experiment import Experiment, add_leaderboard_score
 from nyaggle.experiment.averaging import average_results
-from nyaggle.experiment.gbdt import autoprep_gbdt, experiment, find_best_lgbm_parameter
+from nyaggle.experiment.run import autoprep_gbdt, run_experiment, find_best_lgbm_parameter
